src/requirements.txt
altair==4.2.2
anyio==3.6.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
dulwich==0.21.3
entrypoints==0.4
fastapi==0.95.1
filelock==3.12.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
installer==0.7.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pexpect==4.8.0
Pillow==9.5.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
protobuf==3.20.3
ptyprocess==0.7.0
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
pymongo==4.3.3
Pympler==1.0.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rapidfuzz==2.15.1
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.3.5
scipy==1.10.1
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
streamlit==1.22.0
sympy==1.11.1
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.0
tornado==6.3.1
trove-classifiers==2023.4.29
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.21.1
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
